Autoload customize group
If you don't specify a group key in the globalized minor mode definition, the default group name is the name of the mode parameter minor the "-mode". This seems to be the emacs convention across all of the default customize groups. This PR also autoloads the customize group so the user can turn on global-whitespace-cleanup-mode entirely without having to fiddle with the init file.